Removed delete functions until I can make the process efficient

+++ api/api.ts
@@ -252,7 +252,8 @@ let api = {
         Query.equal("topicId", topicId)
       ]  
     )
-    if(conversations !== null) {
+    if(conversations !== null && conversations !== undefined) {
+      console.log("Conversations not null or undefined", conversations)
       conversations.map( async (convo: any) => {
         await api.provider().database.deleteDocument(Server.conversationsDatabaseID, Server.conversationsCollectionID, convo.$id);
       })
